requirements.txt
absl-py==2.1.0
addict==2.4.0
alembic==1.13.1
appdirs==1.4.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
defusedxml==0.7.1
docker-pycreds==0.4.0
filelock==3.13.1
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git@c023468fafeb0933d06aa8ba49d96089ce9a6af3
grpcio==1.60.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib_metadata==7.1.0
Jinja2==3.1.3
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
lxml==5.1.0
Mako==1.3.5
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.17.3
onnxruntime-gpu==1.18.1
onnxsim==0.4.36
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
optuna==3.6.1
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycocotools==2.0.7
Pygments==2.18.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.39.2
setproctitle==1.3.3
shapely==2.0.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
supervision==0.19.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorrt @ file:///tensorrt-8.6.1-cp310-none-win_amd64.whl
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.12.9
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch @ file:///torch-2.1.2%2Bcu121-cp310-cp310-win_amd64.whl
torchvision==0.16.2+cu121
tqdm==4.66.1
transformers==4.39.2
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.2.17
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.18.1
